+++ src/test/java/com/scanoss/TestScanApi.java
@@ -26,6 +26,7 @@ import com.scanoss.exceptions.ScanApiException;
 import com.scanoss.rest.HttpStatusCode;
 import com.scanoss.rest.ScanApi;
 import lombok.extern.slf4j.Slf4j;
+import okhttp3.OkHttpClient;
 import okhttp3.mockwebserver.MockResponse;
 import okhttp3.mockwebserver.MockWebServer;
 import org.junit.After;
@@ -33,12 +34,13 @@ import org.junit.Before;
 import org.junit.Test;
 
 import java.io.IOException;
-import java.net.http.HttpClient;
+import java.time.Duration;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.concurrent.TimeUnit;
 
 import static com.scanoss.TestConstants.SCAN_RESP_SUCCESS;
+import static com.scanoss.TestConstants.customSelfSignedCertificate;
 import static org.junit.Assert.*;
 
 @Slf4j
@@ -56,9 +58,14 @@ public class TestScanApi {
     }
 
     @After
-    public void Finish() throws IOException {
+    public void Finish() {
         log.info("Shutting down mock server.");
-        server.shutdown(); // Shut down the server. Instances cannot be reused.
+        try {
+            server.close();
+            server.shutdown(); // Shut down the server. Instances cannot be reused.
+        } catch (IOException e) {
+            log.warn("Some issue shutting down mock server: {}", e.getLocalizedMessage());
+        }
     }
 
     @Test
@@ -70,7 +77,7 @@ public class TestScanApi {
         ScanApi scanApi = ScanApi.builder().build();
         assertFalse("Should have pre-configured headers", scanApi.getHeaders().isEmpty());
 
-        scanApi = ScanApi.builder().apiKey("test-key").httpClient(HttpClient.newHttpClient()).build();
+        scanApi = ScanApi.builder().apiKey("test-key").okHttpClient(new OkHttpClient()).build();
         assertFalse("API Key should be set", scanApi.getApiKey().isEmpty());
 
         Map<String, String> headers = new HashMap<>(1);
@@ -82,7 +89,7 @@ public class TestScanApi {
     }
 
     @Test
-    public void TestScanApiScanPositive() throws InterruptedException {
+    public void TestScanApiScanPositive() {
         String methodName = new Object() {
         }.getClass().getEnclosingMethod().getName();
         log.info("<-- Starting {}", methodName);
@@ -101,24 +108,48 @@ public class TestScanApi {
     }
 
     @Test
-    public void TestScanApiScanNegative() throws InterruptedException {
+    public void TestScanApiScanCustomCertPositive() {
         String methodName = new Object() {
         }.getClass().getEnclosingMethod().getName();
         log.info("<-- Starting {}", methodName);
+
+        String sbomIdentify = "pkg:github/scanoss/scanoss.py\n" + "pkg:pypi/scanoss\n";
+        ScanApi scanApi = ScanApi.builder().flags("8").scanType("identify").sbom(sbomIdentify)
+                .url(server.url("/api/scan/direct").toString())
+                .customCert(customSelfSignedCertificate)
+                .build();
+        server.enqueue(new MockResponse().addHeader("Content-Type", "application/json; charset=utf-8")
+                .setBody(SCAN_RESP_SUCCESS).setResponseCode(200));
+        String result = scanApi.scan("file=....", "pkg:github/scanoss/scanoss.py", 1);
+        assertNotNull(result);
+        assertFalse("Should've gotten a response JSON", result.isEmpty());
+        log.info("Scan response: {}", result);
+
+        log.info("Finished {} -->", methodName);
+    }
+
+    @Test
+    public void TestScanApiScanNegative() {
+        String methodName = new Object() {
+        }.getClass().getEnclosingMethod().getName();
+        log.info("<-- Starting {}", methodName);
+        // Empty WFP
         try {
             ScanApi scanApi = ScanApi.builder().build();
             String result = scanApi.scan("", "", 1);
             assertNull("Scan result should be null", result);
-        } catch (ScanApiException | InterruptedException e) {
+        } catch (ScanApiException e) {
             log.info("Got expected Exception: {}", e.getLocalizedMessage());
         }
+        // Invalid URL
         try {
             ScanApi scanApi = ScanApi.builder().url("invalid-url").build();
             String result = scanApi.scan("file=...", "", 1);
             assertNull("Scan result should be null", result);
-        } catch (ScanApiException | InterruptedException e) {
+        } catch (ScanApiException e) {
             log.info("Got expected Exception: {}", e.getLocalizedMessage());
         }
+        // Internal Server Error
         ScanApi scanApi = ScanApi.builder().url(server.url("/api/scan/direct").toString()).build();
 
         server.enqueue(new MockResponse().addHeader("Content-Type", "application/json; charset=utf-8")
@@ -127,36 +158,38 @@ public class TestScanApi {
         String result = scanApi.scan("file=....", "", 1);
         log.info("Scan response: {}", result);
         assertNull("Should've gotten a null response to this scan", result);
-
+        // Service Limit exceeded
+        try {
+            server.enqueue(new MockResponse().addHeader("Content-Type", "application/json; charset=utf-8")
+                    .setBody("Scan failed. Service limit exceeded.").setResponseCode(HttpStatusCode.SERVICE_UNAVAILABLE.getValue())
+            );
+            String result2 = scanApi.scan("file=...", "", 1);
+            assertNull("Scan result should be null", result2);
+        } catch (ScanApiException e) {
+            log.info("Got expected Exception: {}", e.getLocalizedMessage());
+        }
         log.info("Finished {} -->", methodName);
     }
 
-//    @Test
-    public void TestScanApiScanTimeoutNegative() throws InterruptedException {
+    @Test
+    public void TestScanApiScanTimeoutNegative() {
         String methodName = new Object() {
         }.getClass().getEnclosingMethod().getName();
         log.info("<-- Starting {}", methodName);
 
-        ScanApi scanApi = ScanApi.builder().url(server.url("/api/scan/direct").toString()).timeout(2).retryLimit(2).build();
-
-        server.enqueue(new MockResponse().addHeader("Content-Type", "application/json; charset=utf-8")
-                .setBodyDelay(60, TimeUnit.SECONDS)
-                .setBody(SCAN_RESP_SUCCESS).setResponseCode(200));
-        server.enqueue(new MockResponse().addHeader("Content-Type", "application/json; charset=utf-8")
-                .setBodyDelay(60, TimeUnit.SECONDS)
-                .setBody(SCAN_RESP_SUCCESS).setResponseCode(200));
-        server.enqueue(new MockResponse().addHeader("Content-Type", "application/json; charset=utf-8")
-                .setBodyDelay(60, TimeUnit.SECONDS)
-                .setBody(SCAN_RESP_SUCCESS).setResponseCode(200));
-
-//        server.enqueue(new MockResponse().addHeader("Content-Type", "application/json; charset=utf-8")
-//                .setBodyDelay(60, TimeUnit.SECONDS)
-//                .setBody("Scan failed").setResponseCode(HttpStatusCode.INTERNAL_SERVER_ERROR.getValue()));
+        ScanApi scanApi = ScanApi.builder().url(server.url("/api/scan/direct").toString())
+                .timeout(Duration.ofSeconds(2)).retryLimit(2).build();
+        // Setup 3 time-delayed responses
+        for (int i = 0; i < 3; i++) {
+            server.enqueue(new MockResponse().addHeader("Content-Type", "application/json; charset=utf-8")
+                    .setBodyDelay(10, TimeUnit.SECONDS)
+                    .setBody(SCAN_RESP_SUCCESS).setResponseCode(200));
+        }
         try {
             String result = scanApi.scan("file=....", "", 1);
             log.info("Scan response: {}", result);
             assertNull("Should've gotten a null response to this scan", result);
-        } catch(ScanApiException e) {
+        } catch (ScanApiException e) {
             log.info("Got expected Exception: {}", e.getLocalizedMessage());
         }
         log.info("Finished {} -->", methodName);